+++ db/db.go
@@ -6,7 +6,10 @@ import (
 	"global/comp"
 	"os/user"
 	"path/filepath"
+	"strings"
 
+	"github.com/drgrib/alfred"
+	"github.com/drgrib/alfred-bear/db"
 	_ "github.com/mattn/go-sqlite3"
 )
 
@@ -173,4 +176,33 @@ func (lite LiteDB) Query(q string) ([]map[string]string, error) {
 		results = append(results, m)
 	}
 	return results, err
+}
+
+func getUniqueTagString(tagString string) string {
+	tags := strings.Split(tagString, ",")
+	uniqueTags := []string{}
+	for _, t := range tags {
+		isPrefix := false
+		for _, other := range tags {
+			if t != other && strings.HasPrefix(other, t) {
+				isPrefix = true
+				break
+			}
+		}
+		if !isPrefix {
+			uniqueTags = append(uniqueTags, t)
+		}
+	}
+	return "#" + strings.Join(uniqueTags, " #")
+}
+
+func AddNoteRowsToAlfred(rows []map[string]string) {
+	for _, row := range rows {
+		alfred.Add(alfred.Item{
+			Title:    row[db.TitleKey],
+			Subtitle: getUniqueTagString(row[db.TagsKey]),
+			Arg:      row[db.NoteIDKey],
+			Valid:    alfred.Bool(true),
+		})
+	}
 }
